+++ ccip-api-ref/src/components/composed/Callout/index.ts
@@ -1,0 +1,2 @@
+export { Callout } from './Callout.tsx'
+export type { CalloutProps, CalloutType } from './Callout.tsx'